update based on code review comments
