+++ packages/react-core/src/components/NotificationDrawer/NotificationDrawerHeader.tsx
@@ -4,19 +4,15 @@ import { css } from '@patternfly/react-styles';
 import styles from '@patternfly/react-styles/css/components/NotificationDrawer/notification-drawer';
 
 import { Text, TextVariants } from '../Text';
-
-interface NotificationDrawerCloseButtonProps extends React.HTMLProps<HTMLDivElement> {
-  /** Optional icon other than x */
-  icon?: React.ReactNode;
-  /**  Callback for when close button is clicked */
-  onClose: () => void;
-}
+import { NotificationDrawerCloseButton } from './NotificationDrawerHeaderCloseButton';
 
 export interface NotificationDrawerHeaderProps extends React.HTMLProps<HTMLDivElement> {
   /**  Content rendered inside the drawer */
   children?: React.ReactNode;
   /**  Additional classes for notification drawer header. */
   className?: string;
+  /** Adds custom accessible text to the notification drawer close button. */
+  closeButtonAriaLabel?: string;
   /** Optional icon other than x */
   closeIcon?: React.ReactNode;
   /**  Notification drawer heading count */
@@ -31,22 +27,11 @@ export interface NotificationDrawerHeaderProps extends React.HTMLProps<HTMLDivEl
   unreadText?: string;
 }
 
-const NotificationDrawerCloseButton: React.FunctionComponent<NotificationDrawerCloseButtonProps> = ({
-  icon,
-  onClose
-}: NotificationDrawerCloseButtonProps) => (
-  <div>
-    <button className="pf-c-button pf-m-plain" type="button" aria-label="Close" onClick={onClose}>
-      {icon ? icon : <i className="fas fa-times" aria-hidden="true"></i>}
-    </button>
-  </div>
-);
-NotificationDrawerCloseButton.displayName = 'NotificationDrawerCloseButton';
-
 export const NotificationDrawerHeader: React.FunctionComponent<NotificationDrawerHeaderProps> = ({
   children,
   className = '',
   count,
+  closeButtonAriaLabel,
   closeIcon,
   customText,
   onClose,
@@ -64,7 +49,13 @@ export const NotificationDrawerHeader: React.FunctionComponent<NotificationDrawe
     {children && (
       <div className={css(styles.notificationDrawerHeaderAction)}>
         {children}
-        {onClose && <NotificationDrawerCloseButton onClose={onClose} icon={closeIcon} />}
+        {onClose && (
+          <NotificationDrawerCloseButton
+            onClose={onClose}
+            icon={closeIcon}
+            closeButtonAriaLabel={closeButtonAriaLabel}
+          />
+        )}
       </div>
     )}
   </div>
