+++ packages/react-integration/demo-app-ts/src/components/demos/NotificationDrawerDemo/NotificationDrawerGroupsDemo.tsx
@@ -25,6 +25,7 @@ import {
   Title
 } from '@patternfly/react-core';
 import SearchIcon from '@patternfly/react-icons/dist/js/icons/search-icon';
+import CustomCloseIcon from '@patternfly/react-icons/dist/js/icons/outlined-times-circle-icon';
 
 interface GroupsNotificationDrawerDemoState {
   isDrawerOpen: boolean;
@@ -96,7 +97,7 @@ export class GroupsNotificationDrawerDemo extends React.Component<
     ];
     return (
       <NotificationDrawer>
-        <NotificationDrawerHeader count={4} onClose={this.onDrawerClose}>
+        <NotificationDrawerHeader count={4} onClose={this.onDrawerClose} closeIcon={<CustomCloseIcon />}>
           <Dropdown
             onSelect={this.onSelect}
             toggle={
